Calculates the rate of accumulation of phylogenetic diversity (CpD) over time slices

This function estimates the rates of accumulation of phylogenetic diveristy (CpD) over time for inputted assemblages.

Usage

CpD(tree, n, mat, criteria = "my", pDO = 5, ncor = 0)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

n

numeric. A numeric value indicating the number of temporal slices (method = 1) or the time interval in million years (or phylogenetic diversity) among the tree slices (method = 2). Default is 1.

mat

matrix. A presence/absence matrix containing all studied species and sites.

criteria

character string. The method for cutting the tree. It can be either "my" (million years) or "PD" (accumulated phylogenetic diversity). Default is "my".

pDO

numeric. A value indicating the numeric proportion to define the temporal origin at which the phylogenetic diversity (PD) started to accumulate in a given assemblage. Default is 5%.

ncor

numeric. A value indicating the number of cores the user wants to parallelize. Default is 0.

Value

The function returns a data frame containing the assemblages' rates of cumulative phylogenetic diversity (CpD), their total phylogenetic diversity (PD), and their PD origin (pDO).

Details

Parallelization

Users are advised to check the number of available cores within their machines before running parallel programming.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Create a presence-absence matrix
mat <- matrix(sample(c(1,0), 20*10, replace = TRUE), ncol = 20, nrow = 10)
colnames(mat) <- tree$tip.label

# Calculate the CpD for 100 tree slices
CpD(tree, n = 100, mat = mat)
#> > The 100 number of pieces inputted equals to intervals of 0.0243159151955145 million of years.
#>          CpD       PD      pDO
#> 1  0.8796451 8.072499 3.405615
#> 2  0.8345363 6.211377 3.589697
#> 3  0.9817160 9.054914 3.051526
#> 4  0.8339774 7.485656 3.592102
#> 5  0.8537768 7.713385 3.508800
#> 6  0.9364344 8.573761 3.199084
#> 7  0.8395493 7.531502 3.568263
#> 8  0.9362224 6.928680 3.199808
#> 9  1.0036760 9.247250 2.984760
#> 10 0.8488173 7.780532 3.529301