Calculate the tip diversification rates (DR) for a phylogenetic tree

This function computes the tip diversification rates Jetz et al., 2012, or DR, for an inputted ultrametric phylogenetic tree.

Usage

DR(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a data frame containing the tip diversification rates for all species within the inputted phylogenetic tree.

References

See the tutorial on how to use this function on our website. Jetz, Walter, et al. "The global diversity of birds in space and time." Nature 491.7424 (2012): 444-448. doi:10.1038/nature11631

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Computing the tip-DR
DR(tree)
#>    Species       DR
#> 1      t13 2.077467
#> 2      t17 2.077467
#> 3       t9 1.073087
#> 4       t6 4.050308
#> 5      t19 4.050308
#> 6       t4 3.737256
#> 7      t12 3.737256
#> 8      t15 1.221944
#> 9       t2 1.007582
#> 10      t3 1.007582
#> 11      t8 1.314713
#> 12      t7 1.314713
#> 13     t14 3.882436
#> 14      t5 3.882436
#> 15     t11 2.111963
#> 16      t1 6.543343
#> 17     t18 6.543343
#> 18     t16 7.251250
#> 19     t20 7.251250
#> 20     t10 2.183537