Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd  NodeLength YearBegin  YearEnd
#> 22        21      22 1.184166596  0.000000 1.184167
#> 1         22       1 0.868468893  1.184167 2.052635
#> 23        22      23 0.166292108  1.184167 1.350459
#> 24        23      24 0.109762109  1.350459 1.460221
#> 25        24      25 0.329263835  1.460221 1.789485
#> 26        25      26 0.132624799  1.789485 1.922109
#> 38        26      38 0.106300755  1.922109 2.028410
#> 2         38       2 0.024225287  2.028410 2.052635
#> 3         38       3 0.024225287  2.028410 2.052635
#> 4         26       4 0.130526041  1.922109 2.052635
#> 29        25      29 0.173299563  1.789485 1.962784
#> 32        29      32 0.022167523  1.962784 1.984952
#> 33        32      33 0.008317908  1.984952 1.993270
#> 5         33       5 0.059365846  1.993270 2.052635
#> 6         33       6 0.059365846  1.993270 2.052635
#> 36        32      36 0.032337816  1.984952 2.017290
#> 7         36       7 0.035345938  2.017290 2.052635
#> 8         36       8 0.035345938  2.017290 2.052635
#> 35        29      35 0.053926778  1.962784 2.016711
#> 9         35       9 0.035924499  2.016711 2.052635
#> 10        35      10 0.035924499  2.016711 2.052635
#> 30        24      30 0.503234658  1.460221 1.963455
#> 39        30      39 0.077549431  1.963455 2.041005
#> 11        39      11 0.011630586  2.041005 2.052635
#> 12        39      12 0.011630586  2.041005 2.052635
#> 31        30      31 0.010566659  1.963455 1.974022
#> 34        31      34 0.041283421  1.974022 2.015306
#> 13        34      13 0.037329936  2.015306 2.052635
#> 14        34      14 0.037329936  2.015306 2.052635
#> 15        31      15 0.078613357  1.974022 2.052635
#> 27        23      27 0.608523105  1.350459 1.958982
#> 37        27      37 0.059374461  1.958982 2.018356
#> 16        37      16 0.034279219  2.018356 2.052635
#> 17        37      17 0.034279219  2.018356 2.052635
#> 18        27      18 0.093653680  1.958982 2.052635
#> 28        21      28 1.961787243  0.000000 1.961787
#> 19        28      19 0.090848246  1.961787 2.052635
#> 20        28      20 0.090848246  1.961787 2.052635
tree$node_matrix # Node matrix
#>    t20 t1 t7 t19 t4 t14 t6 t18 t13 t11 t9 t10 t15 t12 t2 t3 t16 t8 t17 t5
#> 21   1  1  1   1  1   1  1   1   1   1  1   1   1   1  1  1   1  1   1  1
#> 22   1  1  1   1  1   1  1   1   1   1  1   1   1   1  1  1   1  1   0  0
#> 23   0  1  1   1  1   1  1   1   1   1  1   1   1   1  1  1   1  1   0  0
#> 24   0  1  1   1  1   1  1   1   1   1  1   1   1   1  1  0   0  0   0  0
#> 25   0  1  1   1  1   1  1   1   1   1  0   0   0   0  0  0   0  0   0  0
#> 26   0  1  1   1  0   0  0   0   0   0  0   0   0   0  0  0   0  0   0  0
#> 38   0  1  1   0  0   0  0   0   0   0  0   0   0   0  0  0   0  0   0  0
#> 29   0  0  0   0  1   1  1   1   1   1  0   0   0   0  0  0   0  0   0  0
#> 32   0  0  0   0  1   1  1   1   0   0  0   0   0   0  0  0   0  0   0  0
#> 33   0  0  0   0  1   1  0   0   0   0  0   0   0   0  0  0   0  0   0  0
#> 36   0  0  0   0  0   0  1   1   0   0  0   0   0   0  0  0   0  0   0  0
#> 35   0  0  0   0  0   0  0   0   1   1  0   0   0   0  0  0   0  0   0  0
#> 30   0  0  0   0  0   0  0   0   0   0  1   1   1   1  1  0   0  0   0  0
#> 39   0  0  0   0  0   0  0   0   0   0  1   1   0   0  0  0   0  0   0  0
#> 31   0  0  0   0  0   0  0   0   0   0  0   0   1   1  1  0   0  0   0  0
#> 34   0  0  0   0  0   0  0   0   0   0  0   0   1   1  0  0   0  0   0  0
#> 27   0  0  0   0  0   0  0   0   0   0  0   0   0   0  0  1   1  1   0  0
#> 37   0  0  0   0  0   0  0   0   0   0  0   0   0   0  0  1   1  0   0  0
#> 28   0  0  0   0  0   0  0   0   0   0  0   0   0   0  0  0   0  0   1  1
tree$tree_depth # Tree age
#> [1] 2.052635