Runs a sensitivity analysis for rates of accumulation of a given phylogenetic index

This function allows the evaluation of the sensitivity of the estimated rates of accumulation of a given phylogenetic index (e.g., CpD(), CpE(), CpB(), CpB_RW()) to the number of slices inputted by the user.

Usage

CpR_sensitivity(tree, vec, mat, asb, rate, samp, comp, method, criteria, ncor)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

vec

numeric vector. A numeric vector containing a series of numbers of slices.

mat

matrix. A presence/absence matrix containing all studied species and sites.

asb

matrix, or list of matrices. A matrix (or list of matrices) containing a focal assemblage and its neighborhood assemblages (need at least two assemblages to run).

rate

character string. The desired cumulative phylogenetic rate to be assessed, which can be the phylogenetic diversity (CpD), phylogenetic endemism (CpE), phylogenetic B-diversity (CpB), or phylogenetic B-diversity range-weighted (CpB_RW). Default is NULL, but must be filled with "CpD", "CPE", "CpB_RW", or "CpB".

samp

numeric. The number of assemblages, or sites, to be sampled to make the sensitivity analysis.

comp

character string. The component of beta-diversity that the user wants to calculate the CpB. It can be either "sorensen", turnover" or "nestedness". This argument works only when "rate = CpB". Default is "sorensen".

method

character string. The method for calculating the CpB-rate. It can be either "pairwise" or "multisite". This argument works only when the argument "rate" is set to run for "CpB" or "CpB_RW". Default is "multisite".

criteria

character string. The method for cutting the tree. It can be either "my" (million years) or "PD" (accumulated phylogenetic diversity). Default is "my".

ncor

numeric. A value indicating the number of cores the user wants to parallelize. Default is 0.

Value

This function returns a data frame containing the sensitivity analysis for a given rate of accumulation of a phylogenetic index. This outputted data frame contains, for each row or assemblage, a column with the rate value assessed for each inputted number of slices.

Details

Parallelization

Users are advised to check the number of available cores within their machines before running parallel programming.

Plotting

For plotting the sensitivity analysis output users can use CpR_sensitivity_plot().

References

See the tutorial on how to use this function on our website.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Create a presence-absence matrix
mat <- matrix(sample(c(1,0), 20*10, replace = TRUE), ncol = 20, nrow = 10)
colnames(mat) <- tree$tip.label

# Calculate the CpD for 100 tree slices
CpD(tree, n = 100, mat = mat)
#> > The 100 number of pieces inputted equals to intervals of 0.00996503107798625 million of years.
#>         CpD       PD      pDO
#> 1  2.288306 3.748089 1.309149
#> 2  2.054897 2.507022 1.457850
#> 3  2.330795 3.827401 1.285283
#> 4  2.426811 3.799655 1.234432
#> 5  2.335756 3.776719 1.282553
#> 6  2.256898 3.535351 1.327367
#> 7  2.120633 2.584627 1.412659
#> 8  2.326470 3.639039 1.287673
#> 9  2.212250 3.569210 1.354157
#> 10 2.241953 3.614290 1.336215

# Create a vector of number of slices
vec <- c(25, 50, 75, 100, 125, 150)

# Calculate the sensitivity of the CpD
CpR_sensitivity(tree, vec, mat, rate = "CpD", samp = 5)
#> Warning: Removing the species in presence abscence matrix without any occurrence
#> Warning: Removing tips from phylogeny that are absent on species matrix
#> > The 25 number of pieces inputted equals to intervals of 0.039860124311945 million of years.
#> Warning: Removing the species in presence abscence matrix without any occurrence
#> Warning: Removing tips from phylogeny that are absent on species matrix
#> > The 50 number of pieces inputted equals to intervals of 0.0199300621559725 million of years.
#> Warning: Removing the species in presence abscence matrix without any occurrence
#> Warning: Removing tips from phylogeny that are absent on species matrix
#> > The 75 number of pieces inputted equals to intervals of 0.0132867081039817 million of years.
#> Warning: Removing the species in presence abscence matrix without any occurrence
#> Warning: Removing tips from phylogeny that are absent on species matrix
#> > The 100 number of pieces inputted equals to intervals of 0.00996503107798625 million of years.
#> Warning: Removing the species in presence abscence matrix without any occurrence
#> Warning: Removing tips from phylogeny that are absent on species matrix
#> > The 125 number of pieces inputted equals to intervals of 0.007972024862389 million of years.
#> Warning: Removing the species in presence abscence matrix without any occurrence
#> Warning: Removing tips from phylogeny that are absent on species matrix
#> > The 150 number of pieces inputted equals to intervals of 0.00664335405199083 million of years.
#>         25       50       75      100      125      150
#> 1 2.149513 2.268480 2.309826 2.330795 2.343459 2.351978
#> 2 2.072028 2.183597 2.222352 2.241953 2.253806 2.261761
#> 3 2.044974 2.154719 2.192912 2.212250 2.223946 2.231798
#> 4 1.965192 2.067583 2.102813 2.120633 2.131398 2.138620
#> 5 2.155719 2.273944 2.314959 2.335756 2.348326 2.356767