Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd  NodeLength YearBegin   YearEnd
#> 22        21      22 0.782321637 0.0000000 0.7823216
#> 1         22       1 0.806756858 0.7823216 1.5890785
#> 25        22      25 0.368387957 0.7823216 1.1507096
#> 36        25      36 0.411296473 1.1507096 1.5620061
#> 2         36       2 0.027072427 1.5620061 1.5890785
#> 37        36      37 0.005691771 1.5620061 1.5676978
#> 39        37      39 0.018912743 1.5676978 1.5866106
#> 3         39       3 0.002467912 1.5866106 1.5890785
#> 4         39       4 0.002467912 1.5866106 1.5890785
#> 5         37       5 0.021380656 1.5676978 1.5890785
#> 29        25      29 0.322846369 1.1507096 1.4735560
#> 6         29       6 0.115522531 1.4735560 1.5890785
#> 30        29      30 0.027115311 1.4735560 1.5006713
#> 7         30       7 0.088407221 1.5006713 1.5890785
#> 35        30      35 0.058629316 1.5006713 1.5593006
#> 38        35      38 0.026591742 1.5593006 1.5858923
#> 8         38       8 0.003186162 1.5858923 1.5890785
#> 9         38       9 0.003186162 1.5858923 1.5890785
#> 10        35      10 0.029777904 1.5593006 1.5890785
#> 23        21      23 0.952365992 0.0000000 0.9523660
#> 26        23      26 0.330996550 0.9523660 1.2833625
#> 34        26      34 0.266801388 1.2833625 1.5501639
#> 11        34      11 0.038914565 1.5501639 1.5890785
#> 12        34      12 0.038914565 1.5501639 1.5890785
#> 31        26      31 0.240576892 1.2833625 1.5239394
#> 13        31      13 0.065139061 1.5239394 1.5890785
#> 32        31      32 0.009545148 1.5239394 1.5334846
#> 33        32      33 0.004225020 1.5334846 1.5377096
#> 14        33      14 0.051368892 1.5377096 1.5890785
#> 15        33      15 0.051368892 1.5377096 1.5890785
#> 16        32      16 0.055593912 1.5334846 1.5890785
#> 24        23      24 0.019863923 0.9523660 0.9722299
#> 17        24      17 0.616848579 0.9722299 1.5890785
#> 27        24      27 0.488595940 0.9722299 1.4608259
#> 18        27      18 0.128252639 1.4608259 1.5890785
#> 28        27      28 0.010539360 1.4608259 1.4713652
#> 19        28      19 0.117713279 1.4713652 1.5890785
#> 20        28      20 0.117713279 1.4713652 1.5890785
tree$node_matrix # Node matrix
#>    t7 t11 t18 t16 t10 t1 t6 t12 t19 t15 t5 t20 t8 t3 t2 t4 t9 t17 t13 t14
#> 21  1   1   1   1   1  1  1   1   1   1  1   1  1  1  1  1  1   1   1   1
#> 22  1   1   1   1   1  1  1   1   1   1  0   0  0  0  0  0  0   0   0   0
#> 25  0   1   1   1   1  1  1   1   1   1  0   0  0  0  0  0  0   0   0   0
#> 36  0   1   1   1   1  0  0   0   0   0  0   0  0  0  0  0  0   0   0   0
#> 37  0   0   1   1   1  0  0   0   0   0  0   0  0  0  0  0  0   0   0   0
#> 39  0   0   1   1   0  0  0   0   0   0  0   0  0  0  0  0  0   0   0   0
#> 29  0   0   0   0   0  1  1   1   1   1  0   0  0  0  0  0  0   0   0   0
#> 30  0   0   0   0   0  0  1   1   1   1  0   0  0  0  0  0  0   0   0   0
#> 35  0   0   0   0   0  0  0   1   1   1  0   0  0  0  0  0  0   0   0   0
#> 38  0   0   0   0   0  0  0   1   1   0  0   0  0  0  0  0  0   0   0   0
#> 23  0   0   0   0   0  0  0   0   0   0  1   1  1  1  1  1  1   1   1   1
#> 26  0   0   0   0   0  0  0   0   0   0  1   1  1  1  1  1  0   0   0   0
#> 34  0   0   0   0   0  0  0   0   0   0  1   1  0  0  0  0  0   0   0   0
#> 31  0   0   0   0   0  0  0   0   0   0  0   0  1  1  1  1  0   0   0   0
#> 32  0   0   0   0   0  0  0   0   0   0  0   0  0  1  1  1  0   0   0   0
#> 33  0   0   0   0   0  0  0   0   0   0  0   0  0  1  1  0  0   0   0   0
#> 24  0   0   0   0   0  0  0   0   0   0  0   0  0  0  0  0  1   1   1   1
#> 27  0   0   0   0   0  0  0   0   0   0  0   0  0  0  0  0  0   1   1   1
#> 28  0   0   0   0   0  0  0   0   0   0  0   0  0  0  0  0  0   0   1   1
tree$tree_depth # Tree age
#> [1] 1.589078