Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd NodeLength YearBegin  YearEnd
#> 23        21      23 2.52700096  0.000000 2.527001
#> 35        23      35 0.28742242  2.527001 2.814423
#> 1         35       1 0.08648538  2.814423 2.900909
#> 2         35       2 0.08648538  2.814423 2.900909
#> 27        23      27 0.22338385  2.527001 2.750385
#> 3         27       3 0.15052396  2.750385 2.900909
#> 30        27      30 0.02988789  2.750385 2.780273
#> 4         30       4 0.12063606  2.780273 2.900909
#> 34        30      34 0.02741794  2.780273 2.807691
#> 5         34       5 0.09321812  2.807691 2.900909
#> 6         34       6 0.09321812  2.807691 2.900909
#> 22        21      22 2.11724370  0.000000 2.117244
#> 7         22       7 0.78366507  2.117244 2.900909
#> 24        22      24 0.51587328  2.117244 2.633117
#> 26        24      26 0.11203002  2.633117 2.745147
#> 8         26       8 0.15576177  2.745147 2.900909
#> 28        26      28 0.01461225  2.745147 2.759759
#> 31        28      31 0.02847848  2.759759 2.788238
#> 32        31      32 0.01384789  2.788238 2.802086
#> 38        32      38 0.04706947  2.802086 2.849155
#> 9         38       9 0.05175367  2.849155 2.900909
#> 10        38      10 0.05175367  2.849155 2.900909
#> 11        32      11 0.09882315  2.802086 2.900909
#> 36        31      36 0.04010214  2.788238 2.828340
#> 12        36      12 0.07256890  2.828340 2.900909
#> 13        36      13 0.07256890  2.828340 2.900909
#> 39        28      39 0.12514516  2.759759 2.884904
#> 14        39      14 0.01600436  2.884904 2.900909
#> 15        39      15 0.01600436  2.884904 2.900909
#> 25        24      25 0.06880643  2.633117 2.701923
#> 16        25      16 0.19898536  2.701923 2.900909
#> 29        25      29 0.06891065  2.701923 2.770834
#> 33        29      33 0.03652555  2.770834 2.807360
#> 17        33      17 0.09354915  2.807360 2.900909
#> 37        33      37 0.03645323  2.807360 2.843813
#> 18        37      18 0.05709593  2.843813 2.900909
#> 19        37      19 0.05709593  2.843813 2.900909
#> 20        29      20 0.13007471  2.770834 2.900909
tree$node_matrix # Node matrix
#>    t7 t17 t5 t8 t18 t14 t13 t3 t11 t16 t4 t20 t15 t6 t2 t12 t19 t10 t1 t9
#> 21  1   1  1  1   1   1   1  1   1   1  1   1   1  1  1   1   1   1  1  1
#> 23  1   1  1  1   1   1   0  0   0   0  0   0   0  0  0   0   0   0  0  0
#> 35  1   1  0  0   0   0   0  0   0   0  0   0   0  0  0   0   0   0  0  0
#> 27  0   0  1  1   1   1   0  0   0   0  0   0   0  0  0   0   0   0  0  0
#> 30  0   0  0  1   1   1   0  0   0   0  0   0   0  0  0   0   0   0  0  0
#> 34  0   0  0  0   1   1   0  0   0   0  0   0   0  0  0   0   0   0  0  0
#> 22  0   0  0  0   0   0   1  1   1   1  1   1   1  1  1   1   1   1  1  1
#> 24  0   0  0  0   0   0   0  1   1   1  1   1   1  1  1   1   1   1  1  1
#> 26  0   0  0  0   0   0   0  1   1   1  1   1   1  1  1   0   0   0  0  0
#> 28  0   0  0  0   0   0   0  0   1   1  1   1   1  1  1   0   0   0  0  0
#> 31  0   0  0  0   0   0   0  0   1   1  1   1   1  0  0   0   0   0  0  0
#> 32  0   0  0  0   0   0   0  0   1   1  1   0   0  0  0   0   0   0  0  0
#> 38  0   0  0  0   0   0   0  0   1   1  0   0   0  0  0   0   0   0  0  0
#> 36  0   0  0  0   0   0   0  0   0   0  0   1   1  0  0   0   0   0  0  0
#> 39  0   0  0  0   0   0   0  0   0   0  0   0   0  1  1   0   0   0  0  0
#> 25  0   0  0  0   0   0   0  0   0   0  0   0   0  0  0   1   1   1  1  1
#> 29  0   0  0  0   0   0   0  0   0   0  0   0   0  0  0   0   1   1  1  1
#> 33  0   0  0  0   0   0   0  0   0   0  0   0   0  0  0   0   1   1  1  0
#> 37  0   0  0  0   0   0   0  0   0   0  0   0   0  0  0   0   0   1  1  0
tree$tree_depth # Tree age
#> [1] 2.900909