Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd   NodeLength YearBegin   YearEnd
#> 23        21      23 8.225498e-01 0.0000000 0.8225498
#> 29        23      29 7.085077e-01 0.8225498 1.5310575
#> 36        29      36 5.770198e-02 1.5310575 1.5887595
#> 1         36       1 1.565782e-02 1.5887595 1.6044173
#> 2         36       2 1.565782e-02 1.5887595 1.6044173
#> 30        29      30 6.218748e-03 1.5310575 1.5372763
#> 3         30       3 6.714106e-02 1.5372763 1.6044173
#> 4         30       4 6.714106e-02 1.5372763 1.6044173
#> 24        23      24 4.632098e-01 0.8225498 1.2857596
#> 26        24      26 1.091164e-01 1.2857596 1.3948759
#> 34        26      34 1.658619e-01 1.3948759 1.5607378
#> 35        34      35 2.119511e-02 1.5607378 1.5819329
#> 5         35       5 2.248438e-02 1.5819329 1.6044173
#> 6         35       6 2.248438e-02 1.5819329 1.6044173
#> 7         34       7 4.367949e-02 1.5607378 1.6044173
#> 28        26      28 9.474294e-02 1.3948759 1.4896189
#> 8         28       8 1.147984e-01 1.4896189 1.6044173
#> 9         28       9 1.147984e-01 1.4896189 1.6044173
#> 25        24      25 1.065229e-01 1.2857596 1.3922825
#> 27        25      27 8.505938e-02 1.3922825 1.4773419
#> 10        27      10 1.270755e-01 1.4773419 1.6044173
#> 11        27      11 1.270755e-01 1.4773419 1.6044173
#> 31        25      31 1.555585e-01 1.3922825 1.5478410
#> 33        31      33 1.110743e-02 1.5478410 1.5589484
#> 38        33      38 4.034696e-02 1.5589484 1.5992954
#> 39        38      39 1.987824e-05 1.5992954 1.5993153
#> 12        39      12 5.102037e-03 1.5993153 1.6044173
#> 13        39      13 5.102037e-03 1.5993153 1.6044173
#> 14        38      14 5.121915e-03 1.5992954 1.6044173
#> 15        33      15 4.546888e-02 1.5589484 1.6044173
#> 32        31      32 6.078428e-03 1.5478410 1.5539194
#> 16        32      16 5.049788e-02 1.5539194 1.6044173
#> 37        32      37 4.396456e-02 1.5539194 1.5978840
#> 17        37      17 6.533316e-03 1.5978840 1.6044173
#> 18        37      18 6.533316e-03 1.5978840 1.6044173
#> 22        21      22 5.209844e-01 0.0000000 0.5209844
#> 19        22      19 1.083433e+00 0.5209844 1.6044173
#> 20        22      20 1.083433e+00 0.5209844 1.6044173
tree$node_matrix # Node matrix
#>    t18 t15 t3 t13 t5 t7 t19 t16 t10 t12 t14 t9 t20 t2 t1 t6 t17 t8 t4 t11
#> 21   1   1  1   1  1  1   1   1   1   1   1  1   1  1  1  1   1  1  1   1
#> 23   1   1  1   1  1  1   1   1   1   1   1  1   1  1  1  1   1  1  0   0
#> 29   1   1  1   1  0  0   0   0   0   0   0  0   0  0  0  0   0  0  0   0
#> 36   1   1  0   0  0  0   0   0   0   0   0  0   0  0  0  0   0  0  0   0
#> 30   0   0  1   1  0  0   0   0   0   0   0  0   0  0  0  0   0  0  0   0
#> 24   0   0  0   0  1  1   1   1   1   1   1  1   1  1  1  1   1  1  0   0
#> 26   0   0  0   0  1  1   1   1   1   0   0  0   0  0  0  0   0  0  0   0
#> 34   0   0  0   0  1  1   1   0   0   0   0  0   0  0  0  0   0  0  0   0
#> 35   0   0  0   0  1  1   0   0   0   0   0  0   0  0  0  0   0  0  0   0
#> 28   0   0  0   0  0  0   0   1   1   0   0  0   0  0  0  0   0  0  0   0
#> 25   0   0  0   0  0  0   0   0   0   1   1  1   1  1  1  1   1  1  0   0
#> 27   0   0  0   0  0  0   0   0   0   1   1  0   0  0  0  0   0  0  0   0
#> 31   0   0  0   0  0  0   0   0   0   0   0  1   1  1  1  1   1  1  0   0
#> 33   0   0  0   0  0  0   0   0   0   0   0  1   1  1  1  0   0  0  0   0
#> 38   0   0  0   0  0  0   0   0   0   0   0  1   1  1  0  0   0  0  0   0
#> 39   0   0  0   0  0  0   0   0   0   0   0  1   1  0  0  0   0  0  0   0
#> 32   0   0  0   0  0  0   0   0   0   0   0  0   0  0  0  1   1  1  0   0
#> 37   0   0  0   0  0  0   0   0   0   0   0  0   0  0  0  0   1  1  0   0
#> 22   0   0  0   0  0  0   0   0   0   0   0  0   0  0  0  0   0  0  1   1
tree$tree_depth # Tree age
#> [1] 1.604417