Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd   NodeLength YearBegin   YearEnd
#> 25        21      25 0.9940116041 0.0000000 0.9940116
#> 29        25      29 0.0923982637 0.9940116 1.0864099
#> 30        29      30 0.0073562063 1.0864099 1.0937661
#> 38        30      38 0.0797274347 1.0937661 1.1734935
#> 1         38       1 0.0063025763 1.1734935 1.1797961
#> 2         38       2 0.0063025763 1.1734935 1.1797961
#> 37        30      37 0.0793940449 1.0937661 1.1731601
#> 3         37       3 0.0066359661 1.1731601 1.1797961
#> 4         37       4 0.0066359661 1.1731601 1.1797961
#> 5         29       5 0.0933862174 1.0864099 1.1797961
#> 6         25       6 0.1857844811 0.9940116 1.1797961
#> 22        21      22 0.5486715944 0.0000000 0.5486716
#> 23        22      23 0.1720536600 0.5486716 0.7207253
#> 27        23      27 0.3179039071 0.7207253 1.0386292
#> 28        27      28 0.0345949075 1.0386292 1.0732241
#> 31        28      31 0.0571250553 1.0732241 1.1303491
#> 7         31       7 0.0494469610 1.1303491 1.1797961
#> 32        31      32 0.0006046535 1.1303491 1.1309538
#> 33        32      33 0.0213754144 1.1309538 1.1523292
#> 8         33       8 0.0274668930 1.1523292 1.1797961
#> 9         33       9 0.0274668930 1.1523292 1.1797961
#> 36        32      36 0.0360165692 1.1309538 1.1669703
#> 10        36      10 0.0128257383 1.1669703 1.1797961
#> 39        36      39 0.0096809184 1.1669703 1.1766513
#> 11        39      11 0.0031448199 1.1766513 1.1797961
#> 12        39      12 0.0031448199 1.1766513 1.1797961
#> 13        28      13 0.1065720163 1.0732241 1.1797961
#> 14        27      14 0.1411669237 1.0386292 1.1797961
#> 24        23      24 0.2690605875 0.7207253 0.9897858
#> 26        24      26 0.0339594421 0.9897858 1.0237453
#> 35        26      35 0.1431818208 1.0237453 1.1669271
#> 15        35      15 0.0128689805 1.1669271 1.1797961
#> 16        35      16 0.0128689805 1.1669271 1.1797961
#> 17        26      17 0.1560508013 1.0237453 1.1797961
#> 34        24      34 0.1727670959 0.9897858 1.1625529
#> 18        34      18 0.0172431475 1.1625529 1.1797961
#> 19        34      19 0.0172431475 1.1625529 1.1797961
#> 20        22      20 0.6311244908 0.5486716 1.1797961
tree$node_matrix # Node matrix
#>    t6 t3 t20 t7 t13 t5 t10 t8 t11 t16 t9 t18 t1 t4 t17 t15 t19 t12 t14 t2
#> 21  1  1   1  1   1  1   1  1   1   1  1   1  1  1   1   1   1   1   1  1
#> 25  1  1   1  1   1  1   0  0   0   0  0   0  0  0   0   0   0   0   0  0
#> 29  1  1   1  1   1  0   0  0   0   0  0   0  0  0   0   0   0   0   0  0
#> 30  1  1   1  1   0  0   0  0   0   0  0   0  0  0   0   0   0   0   0  0
#> 38  1  1   0  0   0  0   0  0   0   0  0   0  0  0   0   0   0   0   0  0
#> 37  0  0   1  1   0  0   0  0   0   0  0   0  0  0   0   0   0   0   0  0
#> 22  0  0   0  0   0  0   1  1   1   1  1   1  1  1   1   1   1   1   1  1
#> 23  0  0   0  0   0  0   1  1   1   1  1   1  1  1   1   1   1   1   1  0
#> 27  0  0   0  0   0  0   1  1   1   1  1   1  1  1   0   0   0   0   0  0
#> 28  0  0   0  0   0  0   1  1   1   1  1   1  1  0   0   0   0   0   0  0
#> 31  0  0   0  0   0  0   1  1   1   1  1   1  0  0   0   0   0   0   0  0
#> 32  0  0   0  0   0  0   0  1   1   1  1   1  0  0   0   0   0   0   0  0
#> 33  0  0   0  0   0  0   0  1   1   0  0   0  0  0   0   0   0   0   0  0
#> 36  0  0   0  0   0  0   0  0   0   1  1   1  0  0   0   0   0   0   0  0
#> 39  0  0   0  0   0  0   0  0   0   0  1   1  0  0   0   0   0   0   0  0
#> 24  0  0   0  0   0  0   0  0   0   0  0   0  0  0   1   1   1   1   1  0
#> 26  0  0   0  0   0  0   0  0   0   0  0   0  0  0   1   1   1   0   0  0
#> 35  0  0   0  0   0  0   0  0   0   0  0   0  0  0   1   1   0   0   0  0
#> 34  0  0   0  0   0  0   0  0   0   0  0   0  0  0   0   0   0   1   1  0
tree$tree_depth # Tree age
#> [1] 1.179796