Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd  NodeLength YearBegin   YearEnd
#> 22        21      22 0.617640748 0.0000000 0.6176407
#> 25        22      25 0.157334044 0.6176407 0.7749748
#> 1         25       1 0.204146352 0.7749748 0.9791211
#> 2         25       2 0.204146352 0.7749748 0.9791211
#> 23        22      23 0.018296451 0.6176407 0.6359372
#> 24        23      24 0.069323239 0.6359372 0.7052604
#> 27        24      27 0.192162111 0.7052604 0.8974225
#> 3         27       3 0.081698595 0.8974225 0.9791211
#> 39        27      39 0.078959622 0.8974225 0.9763822
#> 4         39       4 0.002738973 0.9763822 0.9791211
#> 5         39       5 0.002738973 0.9763822 0.9791211
#> 26        24      26 0.129976586 0.7052604 0.8352370
#> 28        26      28 0.066042397 0.8352370 0.9012794
#> 29        28      29 0.007943484 0.9012794 0.9092229
#> 30        29      30 0.015505347 0.9092229 0.9247283
#> 35        30      35 0.016299087 0.9247283 0.9410273
#> 6         35       6 0.038093805 0.9410273 0.9791211
#> 7         35       7 0.038093805 0.9410273 0.9791211
#> 32        30      32 0.008205995 0.9247283 0.9329342
#> 38        32      38 0.036544749 0.9329342 0.9694790
#> 8         38       8 0.009642149 0.9694790 0.9791211
#> 9         38       9 0.009642149 0.9694790 0.9791211
#> 10        32      10 0.046186898 0.9329342 0.9791211
#> 33        29      33 0.024474487 0.9092229 0.9336974
#> 11        33      11 0.045423752 0.9336974 0.9791211
#> 12        33      12 0.045423752 0.9336974 0.9791211
#> 13        28      13 0.077841723 0.9012794 0.9791211
#> 14        26      14 0.143884120 0.8352370 0.9791211
#> 15        23      15 0.343183945 0.6359372 0.9791211
#> 31        21      31 0.931033671 0.0000000 0.9310337
#> 16        31      16 0.048087472 0.9310337 0.9791211
#> 34        31      34 0.004884117 0.9310337 0.9359178
#> 36        34      36 0.007943472 0.9359178 0.9438613
#> 37        36      37 0.010105252 0.9438613 0.9539665
#> 17        37      17 0.025154631 0.9539665 0.9791211
#> 18        37      18 0.025154631 0.9539665 0.9791211
#> 19        36      19 0.035259884 0.9438613 0.9791211
#> 20        34      20 0.043203356 0.9359178 0.9791211
tree$node_matrix # Node matrix
#>    t8 t17 t16 t11 t9 t19 t20 t2 t4 t18 t7 t5 t3 t15 t13 t14 t12 t6 t1 t10
#> 21  1   1   1   1  1   1   1  1  1   1  1  1  1   1   1   1   1  1  1   1
#> 22  1   1   1   1  1   1   1  1  1   1  1  1  1   1   1   0   0  0  0   0
#> 25  1   1   0   0  0   0   0  0  0   0  0  0  0   0   0   0   0  0  0   0
#> 23  0   0   1   1  1   1   1  1  1   1  1  1  1   1   1   0   0  0  0   0
#> 24  0   0   1   1  1   1   1  1  1   1  1  1  1   1   0   0   0  0  0   0
#> 27  0   0   1   1  1   0   0  0  0   0  0  0  0   0   0   0   0  0  0   0
#> 39  0   0   0   1  1   0   0  0  0   0  0  0  0   0   0   0   0  0  0   0
#> 26  0   0   0   0  0   1   1  1  1   1  1  1  1   1   0   0   0  0  0   0
#> 28  0   0   0   0  0   1   1  1  1   1  1  1  1   0   0   0   0  0  0   0
#> 29  0   0   0   0  0   1   1  1  1   1  1  1  0   0   0   0   0  0  0   0
#> 30  0   0   0   0  0   1   1  1  1   1  0  0  0   0   0   0   0  0  0   0
#> 35  0   0   0   0  0   1   1  0  0   0  0  0  0   0   0   0   0  0  0   0
#> 32  0   0   0   0  0   0   0  1  1   1  0  0  0   0   0   0   0  0  0   0
#> 38  0   0   0   0  0   0   0  1  1   0  0  0  0   0   0   0   0  0  0   0
#> 33  0   0   0   0  0   0   0  0  0   0  1  1  0   0   0   0   0  0  0   0
#> 31  0   0   0   0  0   0   0  0  0   0  0  0  0   0   0   1   1  1  1   1
#> 34  0   0   0   0  0   0   0  0  0   0  0  0  0   0   0   0   1  1  1   1
#> 36  0   0   0   0  0   0   0  0  0   0  0  0  0   0   0   0   1  1  1   0
#> 37  0   0   0   0  0   0   0  0  0   0  0  0  0   0   0   0   1  1  0   0
tree$tree_depth # Tree age
#> [1] 0.9791211