Capture general branching information for an inputted phylogenetic tree

This function captures general tree information, including branching, node positions, and depths. It serves as a core function underlying all algorithms for slicing phylogenies.

Usage

nodes_config(tree)

Arguments

tree

phylo. An ultrametric phylogenetic tree in the "phylo" format.

Value

The function returns a phylogenetic tree in the "phylo" format containing three novel pieces of information (stored within "config", "node_matrix", and "tree_depth").

Details

This function captures node and edge information from an ultrametric phylogenetic tree. The function provides a data frame containing detailed branching information (within the internal "config" object), a node matrix (within "node_matrix"), and the tree age (within "tree_depth").

More specifically, the "config" object returns the following information: NodeBegin: the node at which a given branch begins. NodeEnd: the node at which a given branch ends. NodeLength: the branch length of that nodes interval. YearBegin: the year at which a given node begins. YearEnd: the year at which a given node ends.

See also

Author

Matheus Lima de Araujo matheusaraujolima@live.com

Examples

# Generate a random tree
tree <- ape::rcoal(20)

# Capture tree information
tree <- nodes_config(tree)

# Accessing these informations
tree$config # Nodes configurations
#>    NodeBegin NodeEnd NodeLength YearBegin  YearEnd
#> 23        21      23 1.85926003  0.000000 1.859260
#> 26        23      26 0.93018882  1.859260 2.789449
#> 37        26      37 0.26315707  2.789449 3.052606
#> 1         37       1 0.03662016  3.052606 3.089226
#> 2         37       2 0.03662016  3.052606 3.089226
#> 28        26      28 0.07500992  2.789449 2.864459
#> 30        28      30 0.08554116  2.864459 2.950000
#> 39        30      39 0.12865077  2.950000 3.078651
#> 3         39       3 0.01057538  3.078651 3.089226
#> 4         39       4 0.01057538  3.078651 3.089226
#> 5         30       5 0.13922615  2.950000 3.089226
#> 6         28       6 0.22476731  2.864459 3.089226
#> 24        23      24 0.16677663  1.859260 2.026037
#> 25        24      25 0.73511658  2.026037 2.761153
#> 7         25       7 0.32807284  2.761153 3.089226
#> 29        25      29 0.13668840  2.761153 2.897842
#> 35        29      35 0.12201575  2.897842 3.019857
#> 8         35       8 0.06936870  3.019857 3.089226
#> 36        35      36 0.02070094  3.019857 3.040558
#> 9         36       9 0.04866776  3.040558 3.089226
#> 10        36      10 0.04866776  3.040558 3.089226
#> 11        29      11 0.19138445  2.897842 3.089226
#> 27        24      27 0.82771006  2.026037 2.853747
#> 38        27      38 0.22490310  2.853747 3.078650
#> 12        38      12 0.01057625  3.078650 3.089226
#> 13        38      13 0.01057625  3.078650 3.089226
#> 31        27      31 0.09654531  2.853747 2.950292
#> 34        31      34 0.02822403  2.950292 2.978516
#> 14        34      14 0.11071001  2.978516 3.089226
#> 15        34      15 0.11071001  2.978516 3.089226
#> 16        31      16 0.13893404  2.950292 3.089226
#> 22        21      22 1.66502872  0.000000 1.665029
#> 17        22      17 1.42419736  1.665029 3.089226
#> 32        22      32 1.28566522  1.665029 2.950694
#> 18        32      18 0.13853214  2.950694 3.089226
#> 33        32      33 0.02638519  2.950694 2.977079
#> 19        33      19 0.11214696  2.977079 3.089226
#> 20        33      20 0.11214696  2.977079 3.089226
tree$node_matrix # Node matrix
#>    t10 t7 t6 t4 t20 t19 t15 t11 t1 t12 t5 t2 t8 t18 t9 t17 t13 t14 t3 t16
#> 21   1  1  1  1   1   1   1   1  1   1  1  1  1   1  1   1   1   1  1   1
#> 23   1  1  1  1   1   1   1   1  1   1  1  1  1   1  1   1   0   0  0   0
#> 26   1  1  1  1   1   1   0   0  0   0  0  0  0   0  0   0   0   0  0   0
#> 37   1  1  0  0   0   0   0   0  0   0  0  0  0   0  0   0   0   0  0   0
#> 28   0  0  1  1   1   1   0   0  0   0  0  0  0   0  0   0   0   0  0   0
#> 30   0  0  1  1   1   0   0   0  0   0  0  0  0   0  0   0   0   0  0   0
#> 39   0  0  1  1   0   0   0   0  0   0  0  0  0   0  0   0   0   0  0   0
#> 24   0  0  0  0   0   0   1   1  1   1  1  1  1   1  1   1   0   0  0   0
#> 25   0  0  0  0   0   0   1   1  1   1  1  0  0   0  0   0   0   0  0   0
#> 29   0  0  0  0   0   0   0   1  1   1  1  0  0   0  0   0   0   0  0   0
#> 35   0  0  0  0   0   0   0   1  1   1  0  0  0   0  0   0   0   0  0   0
#> 36   0  0  0  0   0   0   0   0  1   1  0  0  0   0  0   0   0   0  0   0
#> 27   0  0  0  0   0   0   0   0  0   0  0  1  1   1  1   1   0   0  0   0
#> 38   0  0  0  0   0   0   0   0  0   0  0  1  1   0  0   0   0   0  0   0
#> 31   0  0  0  0   0   0   0   0  0   0  0  0  0   1  1   1   0   0  0   0
#> 34   0  0  0  0   0   0   0   0  0   0  0  0  0   1  1   0   0   0  0   0
#> 22   0  0  0  0   0   0   0   0  0   0  0  0  0   0  0   0   1   1  1   1
#> 32   0  0  0  0   0   0   0   0  0   0  0  0  0   0  0   0   0   1  1   1
#> 33   0  0  0  0   0   0   0   0  0   0  0  0  0   0  0   0   0   0  1   1
tree$tree_depth # Tree age
#> [1] 3.089226